Added 3 more problems and fixed some warnings

+++ src/euler/level1/Problem036.java
@@ -48,7 +48,7 @@ class Base2Palindromes extends AbstractSequence {
 	}
 
 	@Override
-	public int position() {
+	public long position() {
 		return pos;
 	}
